Created LibraryAppConfig Removed 1 Swagger Dependency #5 Removed * imports Moved LibraryApplication Bean
