+++ library-backend/library-app/src/main/java/org/internship/library/app/LibraryAppConfig.java
@@ -1,0 +1,30 @@
+package org.internship.library.app;
+
+import org.internship.library.api.BookService;
+import org.internship.library.impl.BookServiceImpl;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import springfox.documentation.builders.PathSelectors;
+import springfox.documentation.builders.RequestHandlerSelectors;
+import springfox.documentation.spi.DocumentationType;
+import springfox.documentation.spring.web.plugins.Docket;
+import springfox.documentation.swagger2.annotations.EnableSwagger2;
+
+@Configuration
+@EnableSwagger2
+public class LibraryAppConfig {
+
+    @Bean
+    public BookService bookService() {
+        return new BookServiceImpl();
+    }
+
+    @Bean
+    public Docket api() {
+        return new Docket(DocumentationType.SWAGGER_2)
+                .select()
+                .apis(RequestHandlerSelectors.any())
+                .paths(PathSelectors.any())
+                .build();
+    }
+}
