+++ library-backend/library-app/src/main/java/org/internship/library/app/LibraryApplication.java
@@ -1,36 +1,14 @@
 package org.internship.library.app;
 
-import org.internship.library.api.BookService;
 import org.springframework.boot.SpringApplication;
-import org.internship.library.impl.BookServiceImpl;
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-import springfox.documentation.builders.PathSelectors;
-import springfox.documentation.builders.RequestHandlerSelectors;
-import springfox.documentation.spi.DocumentationType;
-import springfox.documentation.spring.web.plugins.Docket;
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
 
 @SpringBootApplication
-@EnableSwagger2
 public class LibraryApplication {
 
 	public static void main(String[] args) {
 		SpringApplication.run(LibraryApplication.class, args);
 	}
 
-	@Bean
-	public BookService bookService() {
-		return new BookServiceImpl();
-	}
 
-	@Bean
-	public Docket api() {
-		return new Docket(DocumentationType.SWAGGER_2)
-				.select()
-				.apis(RequestHandlerSelectors.any())
-				.paths(PathSelectors.any())
-				.build();
-	}
 }
